+++ tests/Doubles/InitializeCallingExtension.php
@@ -1,0 +1,19 @@
+<?php declare(strict_types = 1);
+
+namespace Tests\OriNette\DI\Doubles;
+
+use Nette\DI\CompilerExtension;
+
+final class InitializeCallingExtension extends CompilerExtension
+{
+
+	public function beforeCompile(): void
+	{
+		parent::beforeCompile();
+
+		$this->initialization->addBody(
+			"\$this->parameters['initializeCallingExtension'] = 'called';"
+		);
+	}
+
+}